refactoring - moved all boostraping logic into same file,
